aula 3 - adicionada funções para garantir que quem está usando as rotas da API é um usuário logado

+++ src/routes/memories.ts
@@ -3,8 +3,16 @@ import { prisma } from "../lib/prisma";
 import { z } from 'zod';
 
 export async function memoriesRoutes(app: FastifyInstance) {
-    app.get("/memories", async () => {
+    // isso é feito antes de qualquer uma das outras requisições abaixo
+    app.addHook('preHandler', async (req) => {
+        await req.jwtVerify();
+    })
+
+    app.get("/memories", async (req) => {
         const memories = await prisma.memory.findMany({
+            where: {
+                userId: req.user.sub,
+            },
             orderBy: {
                 createdAt: 'asc',
             }
@@ -19,7 +27,7 @@ export async function memoriesRoutes(app: FastifyInstance) {
         });
     })
 
-    app.get("/memories/:id", async (req) => {
+    app.get("/memories/:id", async (req, res) => {
 
         // eu espero q o param sempre seja um objeto q contenha um "id" do tipo 
         // string e q seja uuid
@@ -36,6 +44,11 @@ export async function memoriesRoutes(app: FastifyInstance) {
                 id, //msm coisa q id: id
             }
         })
+
+        if (!memory.isPublic && memory.userId != req.user.sub){
+            return res.status(401).send();
+        }
+
         return memory;
     })
 
@@ -54,14 +67,14 @@ export async function memoriesRoutes(app: FastifyInstance) {
                 content,
                 coverUrl,
                 isPublic,
-                userId: '14be3794-8601-4b3c-aded-f50cc111c064' //por enquanto estático
+                userId: req.user.sub,
             }
         })
 
         return memory;
     })
 
-    app.put("/memories/:id", async (req) => {
+    app.put("/memories/:id", async (req, res) => {
         const paramsSchema = z.object({
             id: z.string().uuid(),
         })
@@ -74,7 +87,17 @@ export async function memoriesRoutes(app: FastifyInstance) {
         })
         const { content, coverUrl, isPublic } = bodySchema.parse(req.body);
 
-        const memory = await prisma.memory.update({
+        let memory = await prisma.memory.findUniqueOrThrow({
+            where: {
+                id,
+            }
+        })
+
+        if (memory.userId != req.user.sub){
+            return res.status(401).send();
+        }
+
+        memory = await prisma.memory.update({
             where: {
                 id,
             },
@@ -88,12 +111,22 @@ export async function memoriesRoutes(app: FastifyInstance) {
         return memory;
     })
 
-    app.delete("/memories/:id", async (req) => {
+    app.delete("/memories/:id", async (req, res) => {
         const paramsSchema = z.object({
             id: z.string().uuid(),
         })
 
         const { id } = paramsSchema.parse(req.params);
+
+        const memory = await prisma.memory.findUniqueOrThrow({
+            where: {
+                id,
+            }
+        })
+
+        if (memory.userId != req.user.sub){
+            return res.status(401).send();
+        }
 
         await prisma.memory.delete({
             where: {
